app/requirements.txt
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
cycler==0.10.0
fastapi==0.68.0
h11==0.12.0
imageio==2.9.0
importlib-metadata==4.6.3
kiwisolver==1.3.1
matplotlib==3.4.2
Multi-Template-Matching==1.6.1
nest-asyncio==1.5.1
networkx==2.6.2
numpy==1.21.1
opencv-python-headless==4.5.3.56
pandas==1.3.1
Pillow==8.3.1
pydantic==1.8.2
pyngrok==5.0.5
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.2
scipy==1.7.1
six==1.16.0
starlette==0.14.2
tifffile==2021.7.30
tqdm==4.62.0
typing-extensions==3.10.0.0
uvicorn==0.14.0
zipp==3.5.0

